A sequential search is when you look at each piece of data, one by one, and don’t stop until you find what you are looking for.
A sequential search is when you look at each piece of data, one by one, and don’t stop until you find what you are looking for. You can use a sequential search on any data, whether it is sorted or unsorted (though it would be potentially a slow way to locate what you are looking for if the data is in sorted order!) However, sequential search is the only option you can use when you need to search through data that is unsorted.